:fire: remove blog path
