+++ src/config/path.ts
@@ -1,6 +1,6 @@
 const NAVIGATION_PATH = {
   HOME: '/',
-  BLOG: '/blog',
+  BLOG: '/post',
 } as const
 
 export default NAVIGATION_PATH
