+++ src/pages/post/[...id]/index.tsx
