--- src/templates/home/home-page.tsx
@@ -1,9 +1,0 @@
-import React from 'react'
-
-export default function HomePage() {
-  return (
-    <div className='flex flex-wrap items-start justify-center w-full h-full p-32 text-3xl font-bold break-keep'>
-      안녕하세요, Jay_Log에 방문해주셔서 감사합니다.
-    </div>
-  )
-}
